Disabled single-byte-length string encoding
The C msgpack library seems to have a bug in that it doesn't support the
single-byte length encoding for strings (opcode #xD9). Commented this one
out for the time being, since the raison d'etre of this library is to talk
to an application written in C++ and using the C msgpack lib.